+++ core/model.h
@@ -13,7 +13,7 @@ class Model
 {
     private:
 
-        double scale = 1;
+        float scale = 1;
         Vec3 position;
         bool backfaceCulling = true;
     
@@ -47,7 +47,7 @@ class Model
         Light** lights;
 
         Model();
-        Model(string filePath, Vec3 position = Vec3(0, 0, 0), double scale = 1);
+        Model(string filePath, Vec3 position = Vec3(0, 0, 0), float scale = 1);
         void loadModel(string path);
         void draw(Window &window, Camera *camera);
         void calcular_pontos_3D();
@@ -56,19 +56,19 @@ class Model
         void rotate(int rotacaoX, int rotacaoY, int rotacaoZ);
 
         //position
-        void setPos(double x, double y, double z);
+        void setPos(float x, float y, float z);
         void setPos(Vec3 posicao);
-        void setX(double x);
-        void setY(double y);
-        void setZ(double z);
+        void setX(float x);
+        void setY(float y);
+        void setZ(float z);
         Vec3 getPos();
-        double getX();
-        double getY();
-        double getZ();
+        float getX();
+        float getY();
+        float getZ();
 
         //scale
-        void setScale(double tamanho);
-        double getScale();
+        void setScale(float tamanho);
+        float getScale();
 
         //backface culling
         bool getBackfaceCulling();
