+++ core/vec3.cpp
@@ -9,12 +9,12 @@ Vec3::Vec3():
 {
 }
 
-Vec3::Vec3(double x, double y, double z):
+Vec3::Vec3(float x, float y, float z):
     x{x}, y{y}, z{z}
 {
 }
 
-Vec3::Vec3(double a[], double b[])
+Vec3::Vec3(float a[], float b[])
 {
     x = b[0] - a[0];
     y = b[1] - a[1];
@@ -29,14 +29,14 @@ Vec3::Vec3(const Vec3 &v):
 
 
 
-double Vec3::modulo()
+float Vec3::modulo()
 {
     return sqrt( x*x + y*y + z*z );
 }
 
 Vec3 Vec3::versor()
 {
-    double m = this->modulo();
+    float m = this->modulo();
     if(m == 0.0) return Vec3(0,0,0);
     return Vec3( x / m, y / m, z / m);
 }
@@ -44,16 +44,16 @@ Vec3 Vec3::versor()
 Vec3 Vec3::projetar(Vec3 &v)
 {
     Vec3 tmp{v};
-    double multiplicador = this->produto_escalar(v) / v.produto_escalar(v);
+    float multiplicador = this->produto_escalar(v) / v.produto_escalar(v);
     tmp.x *= multiplicador;
     tmp.y *= multiplicador;
     tmp.z *= multiplicador;
     return tmp;
 }
 
-double Vec3::produto_escalar(const Vec3 &v)
+float Vec3::produto_escalar(const Vec3 &v)
 {
-    double produto_escalar = x*v.x + y*v.y + z*v.z;
+    float produto_escalar = x*v.x + y*v.y + z*v.z;
     return produto_escalar;
 }
 
@@ -66,10 +66,10 @@ Vec3 Vec3::produto_vetorial(const Vec3 &v)
     return tmp;
 }
 
-double Vec3::produto_misto(const Vec3 &v1, const Vec3 &v2)
+float Vec3::produto_misto(const Vec3 &v1, const Vec3 &v2)
 {
     Vec3 tmp = this->produto_vetorial(v1);
-    double produto = tmp.produto_escalar(v2);
+    float produto = tmp.produto_escalar(v2);
     return produto;
 }
 
@@ -78,17 +78,17 @@ bool Vec3::verificar_proporcionalidade(const Vec3 &v)
     return ( ( (x / v.x) == (y / v.y) ) && ( (y / v.y) == (z / v.z) ) );
 }
 
-double Vec3::angulo_entre_vetores(Vec3 &v)
+float Vec3::angulo_entre_vetores(Vec3 &v)
 {
-    double produto_escalar = this->produto_escalar(v);
-    double produto_dos_modulos = this->modulo() * v.modulo();
-    double angulo = acos( produto_escalar / produto_dos_modulos ) * 180 / 3.14159265;
+    float produto_escalar = this->produto_escalar(v);
+    float produto_dos_modulos = this->modulo() * v.modulo();
+    float angulo = acos( produto_escalar / produto_dos_modulos ) * 180 / 3.14159265;
     return angulo;
 }
 
 
 
-Vec3 Vec3::multiEscalar(double escalar)
+Vec3 Vec3::multiEscalar(float escalar)
 {
     Vec3 tmp;
     tmp.x = x*escalar;
